Major refactor of layer configuration code.
This refactoring allows us to add convolutional and pooling layers to a
neural network in addition to fully connected layers.

This introduces a new data structure, layer_t, which describes
all the dimensions and metadata needed to execute each layer of the
neural network. This thus removes the separate num_units, num_rows, and
num_cols arrays.

Weight data is now properly transposed so that the output is the same as
the nontransposed matrix multiply. This caught more bugs in the
transposed matmul...

Matches correct output on nontransposed input for the 3-stage Minerva
nnet.

NOTE: Have not fully implemented conv layers or pooling layers...to come
next.