Use libconfuse config files for declaring network models.
The config file syntax and style is loosely based on Caffe, but
obviously we don't support the same number of features or types of
topologies as Caffe. In particular, we don't support feeding outputs of
layers to more than one subsequent layer (the network must be a simple
linear stack), and we don't (yet) support different activation functions
on each layer.

The FLATTEN and INPUT layer types have been removed, which changes some
logic in how we run the network, but the output is still correct.

Two example network config files have been added.