Move nnet_fwd() code out into a layer instantiation.
The existing nnet_fwd implementation assumes that data is loaded on a
per-layer basis. In order to make this configurable and easier to switch
around, the per-layer implementation is taken out of the main file, and
a compile-time option selects which data-load implementation to use.

For now, per-layer is the only option. Blocked layers (DMA loads/stores
from within each layer) are coming.