Code clean up and API definition.
Each nnet accelerator architecture is defined under layers/ and has to
implement the API described by interface.h. run_layer() is now a common
function used by all architectures. Perhaps in the future we'll allow
this to be overridden.

composable and monolithic have already been done like this (as an
example).

Still have to test that composable and monolithic produce correct values
in simulation.