all: Enable zeropadding of data to byte alignment.
If an alignment is specified in config.h through the DATA_ALIGNMENT
macro, then all rows of input and output data will be zeropadded up to a
multiple of that alignment. Most of the changes are only for re-indexing
of the padded arrays, while the computation doesn't usually have to
change to accomodate additional zeros.

This is needed to more accurately model vector data packing in SMIV and
get better performance estimates by simplifying loop boundary checking,
but SMIV doesn't work just yet.