smiv: Improved validation for SMIV CNN.
We can now pipeline the conv2d_kern_row loop and conv2d_col, which hides
the cost of computing dpN iters and other associated control data.
However, the II of conv2d_kern_row is 4, which is due to two factors:
the initial shift of the shift reg takes 4 cycles, and there is probably
a memory dependence on the psum register file which is preventing the
loop from being pipelined.