smiv: CNN: don't merge and reset psum registers every row.
Instead, just keep accumulating in the same psum registers. It is what
we do in SMIV and it improves performance. Oddly, not inlining
merge_psums here seems to actually improve performance (but just
barely).

The II=4 in conv2d_kern_row is actually due to the loading of new
activations and weights, rather than the shift registers. But, II=1 for
the core datapath loop, so we're good there.