all: Major revamp of layer_t struct.
Instead of conflating the meaning of input/output rows/cols, we now
cleanly separate the definitions of inputs, weights, and outputs.
Naturally, there is some duplication and repetition, but this is not a
huge structure anyways.

We now support data alignment zero-padding on inputs as well as weights.

Passes all tests.