smiv: For CNNs, model real scratchpad sizes and divide up work.
In the past, we assumed we had access to the entire problem in
scratchpad, which is obviously not true. This change sets a hard limit
on the scratchpad size, and we divide up the input problem into
iterations by input channels to fit into the hardware limit.

We also add the UMEMs as a single 3MB block to store input pixels - the
scratchpads are only used for the weights for convs.

There are some limitations:
  - Only support for convolutional layers so far.
  - No support for XY tiling yet (Z tiling only).
  - No more than one extra round of reduction at the very end.

And plenty of code/little hacks/workarounds to clean up...