all: Revamp tracer build system.
Thanks to the LLVM-Tracer change that fixed the call argument forwarding
bug, we don't need to concatenate all the files together anymore. Now we
can do incremental builds, generate LLVM IR per file, and simply link at
the end.

The one remaining nit is that the file containing the main function has
to be compiled and optimized AFTER all other files, because we have to
generate a single final labelmap when we run opt on that module.

We also ensure that the WORKLOAD variable gets set to the right values
automatically when we change compile targets.