smiv: Add a simple analytical performance model.
This handles the basic input cases for each of the accelerator, on a
per-invocation basis. It assumes all data already resides in a local
scratchpad with no data stall cycles other than what is explicitly
accounted for.

For quick verification of Aladdin performance results, this is quite
useful. For reduction and FC, this is very accurate! Convolution may be
off by 10-15% since it's a lot more complicated, but reduction and FC
are probably just off by 5% or less.