smiv: Support different activation functions on the CPU.
When we do this, by default, use ACP and an alternate hw wrapping
function. This different wrapping function uses a different signature
with a array name that is prefixed by acp_, so that we can uniquely
mark any accesses to such a variable as ACP loads and stores in the
config file.

Eventually, we'll want to make this an option to turn on or off in the
network config file directly.