all: Support mixing C and C++ code.
SMIV can now use the Eigen backend, which means we need to change the
compiler from gcc to g++. This is accomplished by adding an additional
file src/arch/smiv.cpp, which does nothing but include src/arch/smiv.c,
so that g++ will accept it as a C++ file.

Refactoring of the makefiles was required to do this. In particular, we
no longer include all the arch files on the command line and use the
ARCHITECTURE macro to disable the ones we don't want. Instead, we
selectively include the right c/cpp file in the Makefile based on ARCH.