scripts: Parse the trace for more accurate cycle counts.
The plotted data now shows cycles from the time the first accelerator is
initiated to the time the last accelerator finishes, which means that
we'll be including more CPU time that accounts for CPU processing.

This is still not ideal though - we should use the profiling
capabilities instead, which will give us even more granular data about
exactly when we start processing a layer.