common: Fix compilation bug introduced by argp.
Compiling with -std=c++11 means strict adherence to the standard is
enforced (__STRICT_ANSI__ is defined). This appears to cause issues with
memset/memcpy when argp is included, possibly because argp disables all
compiler attributes when __STRICT_ANSI__ is defined. memset will hang
when compiled with O2/O3 and segfault when compiled with O0 or O1.

The solution is to set the C++ dialect to gnu++11, which enables GNU
language extensions that may violate the C++ standard.

Change-Id: Ic2dfb92f6877b6537a6cd63faed93b5d0b97fe85