smiv: Swap smiv.c and smiv.cpp.
smiv.cpp used to just #include "smiv.c" and was used so we could include
Eigen when running code on the CPU, but all the actual code was in
smiv.c. This created issues with the Makefiles when compiling with g++
because they would not detect an updated timestamp on smiv.cpp when
changes were made, so it would not try to rebuild.

Swapping these will resolve the problem for g++, but it will simply punt
the problem to the LLVM-Tracer side, which depends on smiv.c (and not
smiv.cpp). For now, this seems to be an acceptable tradeoff.

Change-Id: I92fbc34d5ea4cddbbbbd454cc407910372e7ac81