eigen: Add performance test for convolution.
Eigen is currently about 10x slower than the naive version due to the
shuffling of dimensions that is required. This will be fixed in
upcoming commits -- now that we have an Eigen implementation for every
kernel in the system, we can use Eigen's interface to initialize the
data into the right layouts and dimensinoal orderings from the get-go.

Change-Id: I8d1a68b1bca661a419b0a4af91fa99965826c1f0