eigen: Use Tensor API to initialize data.
Our manual data initialization routines assume a RowMajor storage layout
for almost all data and weights in NCHW format. However, the Tensor
routines require that NCHW corresponds to ColMajor, so in order to use
our original data layout, we needed to do a lot of data transposition,
costing us in performance. By using the native Tensor APIs to initialize
data, we can preserve the logical structure of the data and get the more
optimal data layout.

Change-Id: I5bc6e4fae76cadf99ff6ba32514bae75d67884a9