mkl: Working implementation of 3D convolution.
Currently, we require that the input and output formats of the
convolution are NCHW, even though sometimes other formats (like NHWC)
are more efficient. This means we will always reorder the data after
every layer. In a future commit, we'll relax this requirement to reduce
the number of data reorders.

Change-Id: I2dfbfb2ea92a41afcb446ce06e262a4194484ab4