mkl: Support input/output chaining for FC layers.
This required some changes to BaseMklOp so that all operations can
return the layer configuration. Activation functions are different in
this regard since they don't have one; ideally we would separate this
into two classes (LayerOp and ActivationFuncOp, for example), but this
solution is simpler.

Chaining batch norm and FC layers is weird too because bach norm
requires the dimensions to all be compressed into the channel dimension,
which messes up reordering (the dims have to be the same), so we bypass
chaining for BN->FC.

Change-Id: I9928be376ff27a08aba82a9a5fcd35b125551c4d