mkl: Fix invalid reference segfaults to memory primitives.
Simply storing the mkldnn::memory primitive objects in the memories
vector does not work because when the vector needs to be resized, all
objects in the current vector are copied into the new resized vector,
which means that all references to the original object are now stale and
invalid.

The solution is to simply store a pointer in the vector instead. We use
unique pointers to deal with management.

Change-Id: I297839a02a6d344454b4d3024fdefdef437899c9