smiv: Support work division for large FC layers.
When an FC layer is too large for all the weights to fit into UMEM, the
weights are divided up into chunks that are individually computed. Each
chunk is num_inputs x num_chunk_neurons, where num_chunk_neurons must be
a multiple of 8 (the SMIV vector size). Results are sent back to the
host after every iteration and aggregated at the end, with some
reordering required if the batch size is greater than 1.

Change-Id: Iaffba0fc0e04738dfda83837995c40d8a8f99904