smiv: By default, use the reference impl on the CPU for batch norm.
SMIV doesn't support batch norm in HW anyways, so unless we want to
model it explicitly, use the software fallback. Eventually, we'll want
to use MKL for BN, but we need to implement one that supports
precomputing 1/sqrt(var).

This also fixes a bug where results of BN layers were not getting DMAed
back to the host.

Change-Id: Ideab8591a7d3eaf0737444cf9560738870e7667b