smiv: Support depthwise convolutions.
This support is incomplete though, because SMIV doesn't support
activation functions on the CNN block (it does it on the reduction
block). We'll need to add support for this, even though it doesn't exist
on the real HW. For now, we run all activation functions on the CPU
(including RELU).

Change-Id: Icf19185d3600a5794a3284734a4b2794ca6ca11a