all: Small cleanups.
In particular, this adds a small function to compute the size of a
dims_t object. This pattern gets used a lot in this codebase, but it's
usually manually written out with layers[lnum].inputs... because of
Aladdin restrictions on array base addresses. This function can be used
freely on the CPU but it should be used with care in an accelerated
section.

Change-Id: I81c9af5da6eda1a36f68138fc6099e1f3e8759b3