mkl: Give the layer descriptor to activation functions too.
This is required for some kernel to chain itself after an activation
function. In the past, this issue did not crop up because we always put
a pooling layer between CONV/FC layers, and pooling layers don't usually
have activation functions. But for CONV/FC -> CONV/FC to work, this is
required.

Change-Id: Ib2d500d140723cca6d9834a50f2ee137d78b4c59