core: Fix striding bugs in reference and SMIV convolution.
The reference implementation was committing the finished sums to the
wrong place because it didn't track the output pixel row/col separately
from the input pixel row/cow.

The SMIV implementation had a bug in the SIMD implementation (not the
scalar implementation) where finished output pixels would get
overwritten by the results of the next column.

Change-Id: I8c184b8d6d03e32ece5d433a5c428183de4720f8