smiv: Fix data alignment when using the MKL pooling operation.
MKL doesn't support data alignments like we do, but we can fake that by
introducing additional padding on the right side of each row of the
input image. This patch calculates the appropriate amount of padding
required based on the input/output rows and pooling size/strides. The
equation could likely be simplified.

Change-Id: I95b1eddcd541045a68d331fe3f47156b6fcbc55f