smiv: Support sampled execution for standard convolutional layers.
This is a very coarse sampling strategy that reduces the total number of
filters that are simulated per layer. The idea is that each output
channel takes roughly the same amount of time to run (assuming little
shared resource contention), which means this is a good candidate for
a simple sampling technique.

Change-Id: I5802499d7ce1ef6325b5a503bed7aa9e7e3359cb