smiv: Implement a row-wise division of inner product weights.
The existing work division scheme works by splitting the FC weights by
groups of COLUMNS. This requires us to reorder the FC weights, which we
have observed to be very expensive.

An alternative way is to split the FC weights by ROWS, which allows us
to just pass a pointer to the weights, no reordering needed. However,
the INPUTS will need special handling, since we're only using a fraction
of them at once. They require an additional round of reduction at the
very end and some minor shifting as well. Furthermore, we can no longer
run supported activation functions on the HW, since we never have the
fully reduced sums in HW. Finally, we add the ability to ignore biases,
since they can't be added until the very last iteration. This ability is
present in the real HW anyways.

The hope is that the cost of doing the additional input reduction and
software activation functions will outweight the cost of the FC weight
reordering when the inputs are small compared to the weights. If the
reverse is true, then we will still use the original columnwise work
division.

This is a big commit that does a significant amount of refactoring to
the inner product code to maximize code reuse.

Change-Id: Ia1c0102d00db3f43ccbb77bac28979adae555538