core: Add a layer number field to the layer_t struct.
Not having this causes us no end of trouble because of the need to pass
around and plumb through an extra parameter to tons of function calls.
I avoided this change for awhile because there is so much existing code
that uses the extra lnum/layer_num parameter, because it would require
changing all the Aladdin configuration files to update the layer_t
struct size, and the potential for causing memory dependence analysis
issues. At this point though, avoiding it is simply not worth it
anymore.

The remainder of SMAUG will get refactored over time to use this field.

Change-Id: Ie4a00754df741dbd46b8147b90d19387a96d134e