utility: Add an API to start an ignored profiling section.
Ignored sections are labeled with __IGNORE__. This can be used to remove
certain sections that might contribute overhead we don't want to
measure.

Note that such a section does not actually subtract its elapsed time
from its parents. The profiling log is unchanged. It is the
responsibility of consumers of the profiling log to take this into
account.

Change-Id: Iae2b636dd06ed5f5d27fcb24a0d2de1919a34f98