all: Enable batch norm in HW, if the HW supports it.
Like activation functions, this can be switched on and off with a
configuration file setting in the device setting. By default, this is
set to false.

Currently the HW performance of this implementation is quite slow due to
bank conflicts when accessing the various parameters required for BN.
This can only be improved by using a different data layout, which will
most likely be done at runtime instead of at initialization, because
every other backend uses the current blocked BN weight format.

Change-Id: I590d356c5c1650069348ce9ccc8d8aca950eb72a