mkl: Implement a batch norm op that uses precomputed weights.
This will replace the existing MKL batch norm, which still has to do a
division + sqrt and appears to contribute significantly to cycles spent
on MobileNets. The original MKL implementation is still there, of
course.

Some basic performance testing shows about a 10x speedup over MKL,
mostly because MKL was doing an expensive scalar divide and multiply on
every iteration.

Change-Id: Id7077767059d18fe5632211548e27b8d1dedff12