smiv: Implementation of max pooling in HW.
The pooling kernel works on blocked NHWC data, unlike the convolutional
block, which requres NCHW data. As expected, this was much simpler to
implement. We support multiple iterations of pooling as well. This
implementation does not account for the activation function.

The current dataflow just DMAs input data to SPAD0 and output data from
SPAD1. The UMEM is not used at all. This can be refined in the future by
using the UMEM to store all the input data and manually pulling data out
of the UMEM into SPAD0/SPAD1, so that we can do full double-buffering all
on the hardware without any more CPU intervention.

Change-Id: I5ddab693c4205e93b0f24c2f1c126803e4a96fd5