smiv: Reuse inputs to convolutional layers across output channels.
We only need to DMA the input image to the UMEM once. There's no need to
do this once per kernel, since the UMEM *must* fit the entire image,
and we don't need to reuse the UMEM storage for any intermediate outputs
at this point either. This should reduce DMA time on convolution
significantly.

A few things needed to be changed:
  1. Reduction writes to a scratchpad instead of the UMEM. This is fine
     since we're guaranteed to fit the reduced results into the
     sratchpad.
  2. Convolutions had to be modified to start reading input not from the
     base of the inputs memory, but from the current iteration's
     starting channel instead.

One note: when a convolutional layer takes multiple iterations per
output channel, one optimization is to store the partially reduced sums
back to the UMEM, so that we don't need a DMA on the final reduction.
This change would need to be revisited for that to happen, but the wins
from this commit should easily outweigh storing partial sums in the
UMEM.

Change-Id: Ia58118dbca2626a2e01c92387204533d745445a0