core: Improve the LUT sigmoid implementation.
The existing implementation uses a centered sigmoid table (where x
ranges from -N to N) and linearly interpolates between the points.
Instead, we can store a non-centered table (from 0 to N) with twice the
precision for the same table size. With a large sized table and twice
the precision, linear interpolation becomes less necessary.

This commit leaves both implementations in place. The user can decide
which one to use with a new command-line argument: --sigmoid-impl or -m.

Change-Id: I80c63d7e250d384cf83e7a11c4e7eecbd6fcfb26