mkl: Implement LUT-based activation functions involving exp.
This is a tradeoff between getting vectorization with the MKL eltwise
primitives and using a faster per-element lookup function. To use these
primitives instead of the MKL-DNN primitives, pass the -m <option> flag,
just as before.

Based on some simple performance tests, the noncentered LUT sigmoid
implementation is faster than MKL-DNN, but ELU and SELU are about the
same or slower.

Change-Id: Ie5d783627382331e5ef1156958233a70bdf00c35