smiv: Add all-cache, all-acp support for the accelerator.
We only used ACP for transfering the results. This adds the support
for using cache and ACP for transfering everything (activations,
weights and results). Specifically,

1) for the inner product block, this adds cache-only, acp-only options
for accelerator memory interface design. The scrachpads and shared umem
are not used when acp/cache is enabled.

TODO: when using cache/acp for everything, we do not need to do work
division anymore, since we are not using the scratchpads/umem.

2) For the convolutional blocks, it's a bit different because we still
use scratchpads for storing the convolution results. Specifically, the
standard convolutional block outputs the results to the local scratchpad
and the reduction block reads directly from it. The cache-only, ACP-only
support will not change this. Other than that, the activations/weights
of the convolutional blocks and the result of the reduction block all
use ACP/cache.

This also adds pipelined DMA optimization for FC and Conv blocks.

Change-Id: Id773e021613b2b1109c6ba8d2621b74bb60a9248