smiv: Support activation functions for 2D convolutions.
This runs the activation function on the completed partial sums right
before they are committed. For lookup-based activation functions, this
should only take 2-3 additional cycles (assuming no bank conflicts on
the inputs), and the outer loop is pipelined anyways.

For standard convolutions, this will not do anything - the activation
function is still executed by the reduction. This only applies for
depthwise convolutions.

Added a new test for this.

Change-Id: Ia4d76f6a84f2a12647ddfc3548e502e7552cda6c