utility: Support reading a layer-wise compression mask from saved data.
Currently we don't save data in compressed form, so in order to support
sparse weights we need the framework to know whether or not to compress
weights manually before starting the forward pass.  This mask indicates
whether a layer's weights should be compressed into sparse format after
being read out of a data archive.

This supports both text and binary formats, both reading and saving.

Change-Id: I6550e0d0367bef3c31d68c2b1059223f361c417a