all: Integrate CSR compression into main binary.
If a COMPRESS TYPE section is present in a txt model archive, and the
compression masks indicate that certain layers should be compressed,
then the main binary will compress them into the modified CSR format and
attach them to the layer_t struct.

For now, no backends actually take advantage of the compressed data;
that will be coming up in the next commits.

Change-Id: If7dccc347de693ed82c2c76313e93e3a8f4c4c5f