utility: Fix bug in reading sections from binary data archives.
We used to update the file->addr pointer after reading a section
header/section payload to mimic the side effects that IO function calls
have on file descriptors/FILE pointers, but that makes it harder (and
would break if we tried) to read sections out of order. There's really
no reason to have this feature. Don't update this pointer and always
seek from the start of the file.

Change-Id: Ief399e7b9bcacde9ac9262f4ae16e63ea0c9e2cb