all: Add a CLI arg to convert between txt and bin data formats.
This lets us produce a binary data archive given a text archive and vice
versa. Simply add the "-c" flag. If this flag is specified without a
data file name, then it is ignored. After conversion, the program will
stop; it is assumed that the user only wants to do the conversion.

Change-Id: I366b9944b5764576fe6877914fadedd11d18e08b