smiv: Add HW support for decompression of CSR weights.
This is implemented as a separate accelerator block that runs just prior
to starting an FC layer. The CSR decompression block reads in the packed
inputs and indices into one of the scratchpads, zeros out the part of
the UMEM it will need to modify, and writes the decompressed data into
the UMEM. Then, the FC block can start to run without needing to load
the weights.

Change-Id: I67d180c01881c5df046df45e145c13f4ddf39f9c