common: Get rid of some spurious compiler warnings.
Any variable used in a MAP_ARRAY_TO_ACCEL call would trigger an
unused variable warning for non-gem5 builds, so make the macro sink the
variable instead.

Also, a gcc 5.4.0 bug seems to cause a ton of spurious maybe-initialized
warnings with <anonymous> as the source inside one file, so disable that
warning in that file.

Change-Id: I5ba9543e4f552f505ff55fab46afea8f60f6d125