smiv: Support tiling CSR if too large to fit in scratchpad.
Tiling a sparse data structure is extremely annoying because we have to
actually ascertain the sparsity per row/column before determining
whether adding it would require a new tile. This commit adds support for
breaking up a large CSR into multiple tiles, based on the scratchpad
size, and decompressing the tiles one at a time into the UMEM.

The "optimal" way to tile this depends on both the compressed and
uncompressed array sizes. To simplify matters,  this implementation
first tiles the uncompressed data and then tiles that corresponding
section of the sparse array.

This commit also fixes several bugs that were present in the initial CSR
implementation and a few more mapping related bugs.

Change-Id: I13c67d58c1d29a3db91df52804a8b90b89790de0