models, tests: Add a real-data test for CIFAR10 and update expected results.
This fixes a few bugs with the model - we specified the wrong number of
output channels for the later convolutional layers and the wrong number
of input channels.

The old random-data test for CIFAR10 used a single input channel instead
of three. And now with the fixes to CSR, using CSR or not using CSR
produce results that are much closer.

For some reason, the composable test is broken on CIFAR10. I'm going to
leave it broken for now though, since we don't have time to fix it.

Change-Id: Ica18d8fa0748cd8580426f40953738ec1ee92a32