smiv: Fix strange bug induced by gcc optimizations.
To make indexing into a flat array easier, we cast it into a pointer-
to-array, so that we can use the pointer as if we are using a
multi-dimensional array. However, gcc apparently does unwanted
optimizations when we use local const variables as the dimension
lengths of the ND-array. Don't know the exact cause of it though.
This removes all the const keywords. After the change, all the all
the "<anonymous> maybe-uninitialized" warnings are gone, so re-enable
those warnings in convolution.c. They can be good indicators if this
bug strikes back in the future.

Also, this fixes a few mapping bugs in invoking the reduction block.

Change-Id: I29d7ce4b354181cdc81d898722ec3c7d638670e4