common: Fix bug with missing labels in the labelmap.
The instrumentation pass must not run on the file containing the main
function until the very end, because when it sees the main function, it
will immediately insert the contents of the labelmap as it currently
stands on disk into the binary. We need to wait for all other functions
first before doing this.

Technically, we should just not run the labelmapwriter opt pass at all
until the very end. That'l come in a future patch.

Change-Id: I86c24d3bd9a71090a28a39dadf46dc77bb7e03ae