smiv: Unify the dispatch mechanism for inner product.
In the past we had to create different functions for different
combinations using DMA/ACP/Cache for inputs/weights/outputs, which was
fine at first but quickly became unscalable.

This commit creates a single hardware entry point and contains a
function argument for each array and access mechanism combination
(dma_weights, cache_results, etc). It then calls the underlying
implementation with the appropriately named argument.  Code repetition
is reduced with the use of macros that allow us to specify just the
mechanism in place of the complete variable name (in many cases).
Additional complexity with whether to use spad0 or spad1 is handled by
another macro that acts as a mux.

Change-Id: I58d69c9445609049c09a8d1dc04ce842c1f44155