smiv: Unify access mechanisms for convolution + reduction.
This works just like the past refactor for inner products. We
consolidate all of the individual functions that had to be dispatched
conditional into a single function which runs in the HW domain.

All existing tests are passing. More tests will be added to test
different combinations of access mechanisms.

Change-Id: I07ec13e4a835e62355f9aba214c5960247f19541