smv: DLA-like convolutional core kernel implementation.
Unlike SMIV, this convolutional kernel operates on NHWC data, which is a
far better design for image classification workloads. This
implementation uses 4-sets of 8-way MACCs per PE. Each PE operates on a
distinct output channel and performs a 32-way spatial reduction on each
input channel. We have 4 PEs in this design.

The output is written in NCHW layout, because this implementation does
not use vector literals and so it can do arbitrary 32-bit stores to the
output scratchpad.

Some cosmetic changes from Sam were made.

Change-Id: Ia53843b0f4e5d6e7a19627970b207761e6ed375d