smiv: First pass at integrating the new convolutional kernel.
This accelerator model more properly belongs to a new backend (SMV), but
in the interest of time, we'll just insert it in place of the current
SMIV backend.

This is a very rough integration pass, so a lot of features are not
supported, such as input tiling, different access mechanisms, and
activation functions in HW, etc.

Change-Id: I933e52a5a603ac17aa52270827f697692629d378