smv: Rewrite the core convolutional kernel loop nests.
The original code was not actually reusing the weights for a given input
window -- instead, it was accessing the weights scratchpad on every
iteration of the innermost loop. This commit ensures that we can reuse
inputs AND weights as much as possible in this dataflow.

The original code also separated register reset from load to avoid
putting a conditional in the innermost loop, which I thought would be
bad for performance. It turns out that Aladdin handles this just fine,
as long as it is a simple expression.

Performance model under Aladdin on a simple 4x4x32 input estimates 188
cycles, compared to 144 assuming 100% throughput. The difference comes
from just loading the kernel register from the scratchpad prior to
starting the main input window loops, which takes 2 cycles.

Change-Id: I0ef04c0df67436f15382da0aabf9e1347ac0c961