smv: SIMD implementation of the new conv kernel.
The kernel stores the output data in NCHW format, so all partial sums
get buffered in a 8-wide vector object and stored as a group of 8
columns in order to enforce the property that the arrays never get
dereferenced as a scalar. In order to write the data in NHWC format, it
would be best to increase NUM_PE_INSTS from 4 to 8 as so to match the
vector width.

Change-Id: Icebcdadaaada6efcbd2ccf848a991454473a4083