smv: Add sampling support for convolution.
This sampling parameter is very low-level - it essentially operates over
groups of output feature maps, where the group size is the number of
PEs. We ensure we always capture groups that are performing relevant
data movement, but apart from those, we can skip most of the rest.

As an example, suppose we are operating on 32x32x32 input image,
producing 32 output feature maps. This requires one tile and eight
iterations on the HW, each iteration producing four output feature maps.
Those iterations would look like so:

   1. DMA all inputs, DMA all weights, produce ofmaps 0-3.
   2. Produce ofmaps 4-7.    _ _
   3. Produce ofmaps 8-11.    |
   4. Produce ofmaps 12-15.   |
   5. Produce ofmaps 16-19.   |  "Inner iterations"
   6. Produce ofmaps 20-23.   |
   7. Produce ofmaps 24-27.  _ _
   8. Produce ofmaps 28-31, DMA all outputs back.

The sampling parameter controls how many of these inner iterations are
executed. The first and last are always executed. If the value is 0,
then sampling is disabled. This is also on a per-tile basis.

Change-Id: If8f267ecae5686870ab4d9ac4099c40b2d69bf14