utility: Fix col-major FC weights initialization with padding.
When TRANSPOSE_WEIGHTS = 1, we store FC weights in col-major order,
which means that the *rows* are now the innermost dimension and
therefore require data alignment padding. However, layer.weights, for
now, is still going to store the logical dimensions of the weights,
where rows represent input neurons + 1 for the bias and cols represent
output neurons.

These changes fix weight initialization and size calculations to ensure
that the correct sizes and dimensions are calculated. This is really
hacky of course. We need to address the big problem of data layouts and
dimensions properly when we have time.

Change-Id: If948832c639f3863cc192e6b603dda6f2a755567