smv: New matrix multiply kernel with 4-PE 32-way datapath.
This kernel is designed to use the same dataflow as the convolutional
block, so they have the same performance characteristics, but keeping
them as separate implementations makes it easier to manage different
tiling behavior.

The dataflow is output-stationary within a PE. Each PE works on a
different neuron and performs a 32-way vector dot product between the
input and weights for that neuron. It continues in groups of 32 until
the neuron is completely computed. The design has the same number of PEs
as the convolutional block.

Like the convolutional kernel, the bias and activation function are
added when everything is finished, not inside this kernel.

Change-Id: I6f93040937c48c0cbe126b3960682314e28c3d0b