smv: Add software tiling wrappers for the new FC block.
Unlike tiling in SMIV, we actually do not need to rearrange any data as
long as the weights arrive in transposed (col-major) format, because the
design requires the complete input and complete set of weights per
neuron. All that's required is to update the weights pointer and
starting output neuron of each tile.  When the matrix multiply is
finished, we then add the bias and activation function.

Because the weights must be transposed, if weights need to be
compressed, then we have to make sure we compress the transposed
weights. This is not yet supported.

Change-Id: I834be5b5e0ec37dd8de6945faad147b50a500669