arch: Remove redundant dmaStores from the end of nnet_fwd.
This was required only for monolithic originally, not for any of the
others, but it made its way into the others thanks to just copying the
nnet_fwd function around. Monolithic technically doesn't work with
Aladdin anymore, but we'll leave this part in it in case we decide to
make it Aladdin-compatible again in the future.

Change-Id: I333010055b56809188c2c7825be40f69d48fc026