all: Specify bias dimensions independently of other weights.
Since the beginning of SMAUG, biases for FC layers were stored as an
additional row to the weights, which allowed us to re-use the weights
dims_t struct to also encompass the biases. However, this led to two
problems: we could not specify biases for convolutions, and we could not
support col-major FC weights, in both cases because there was no field
to store the alignment padding required for the biases.

This patch adds a bias dims_t struct specifically to store the bias
dimensions and removes the extra +1 in the weights rows. However, most
of the current matrix multiply kernels fuse the GEMM with the bias and
assume an additional row in the weights. To minimize code changes for
the backends that use these fused kernels, I simply compensated by
adding the bias row back in the tiling code. However, backends that
use transposed FC weights with data alignment requirements can now
manipulate biases independently of the rest of the weights.

All the functions that computed the number of weights for each layer had
to be updated to account for the biases being specified separately.

Change-Id: I7f2911ff6176e9ae3b9192970e2b4f823095aa40