smiv, smv: Consolidate all global state into a common interface.
The smiv_global and smv_global structs are used to describe this
interface. By using common names for different accelerator ids, we can
use SMIV hardware blocks from SMV and vice versa. In the past, this
cross-backend usage would cause linking errors because the global
constants were being defined in a place that was not being compiled.

These structs are currently identical so that they can be safely casted
back and forth, but that doesn't mean they must continue to be
identical.

Change-Id: Iad71b99adea5c630d2bdb81aae8f5c1943b0249c