smv: Rewrite 32-way reduction for more efficient graphs.
The old way to write this reduction was to individually add each scalar
element of each vector. This is inefficient because each such scalar
access generates an ExtractElement operation and a ton of address
calculations. Instead, we can use four vector-wide adds to accumulate
down to just a single vector and then do a 8-way reduction, which
reduces the number of scalar accesses.

On some basic testing, this change reduced the node count in the new
matrix multiply and convolution routines by 2x!

Change-Id: I7509e547076c91f04c62691314efc313a86f3a9f