smv: Work around an Aladdin dependence analysis bug.
This bug caused InsertElement operations on a vector register to be
serialized even if they were going to different locations. The
serialization was introduced in the phi node removal pass and is due to
a loop-carried dependence on the partial sums register in the outer loop
weight column loop.

To work around this problem, just convert the vector register into a
normal scalar array, which have much more mature support in Aladdin.

Change-Id: I2ced7b389bbb8893dfbd6a4e4d60a0a7918aa748