smv: Generalize the per-layer weights buffers.
When we introduced support for compressed weights, we added a field to
the layer_t struct to store a pointer to the weights for that layer.
Those weights could be either uncompressed or compressed, but it had to
be a single 2D block of memory. This means that we are unable to support
col-major FC weights when the layer requires a bias, because their
padding requirements may be different from the biases' padding
requirements and thus we cannot compress the bias together with the
weights in a single structure.

This change generalizes this field into a list, so that a layer can
independently store distinct types of weights. Each entry in the list
could in principle be stored in a different format -- weights are
compressed, biases are uncompressed, etc.

This change does not actually resolve the bias compression problem. That
will be next.

Change-Id: I70a121c4e89966c5a2c6bccaa764dba03b780ab7