smv: Add a standalone block just for DMA operations.
This block can be used to invoke the DMA operations from the CPU. This
is useful if we need to do a one-off transfer without changing the API
of an accelerator.  We can make a call to this function "belong" to any
accelerator by specifying the appropriate accelerator id.

This block does not support using the pipelined DMA wrappers due to
current limitations in the DDDG analysis with identifying the correct
base address, although it could be worked around.

Change-Id: Id554467475a65eaba715e939aa4ea741d44b2095