smv: Fix handling of tiled FC with compressed weights.
When weights are uncompressed, we can afford to keep the partial sums
from previous tiles in the scratchpad because we are careful not to
overwrite them on future tiles. Then, on the last tile, we add the bias,
run the activation functions, and send all the data back.  The problem
with doing this when weights are compressed is that we will overwrite at
least one of the scratchpads when reading in the compressed weights
data. We overwrite either the spad holding the input activations or the
spad holding the output activations, so either way, we have to re-copy
something over.

We choose to DMA the outputs after every tile back to the CPU, because
outputs are generally smaller than inputs. On the last iteration, just
before starting the FC, we send all the partial sums back to the
scratchpad, so that the FC will finish up the last tile of GEMM and run
the bias and activation function as usual.

Note that this is only a problem with using DMA. If we were using ACP or
caches, then the host memory will get updated correctly and we don't
need to worry about this.

Change-Id: I03bbe6223cb54d100267848a4d926c99ffb9da50